Unify MCModel to either return as-is or logsoftmax
MCModel now returns either model(x) or logsoftmax as an
an additional layer in all 3 functions: forward, stochastic_forward, and
predict. The acquisition functions are adjusted to account for this
change. ALRModel.eval can now consistently calculate losses without an
additional forward pass since the criterion for predict is similar to
forward.